devtools/server/tests/chrome/test_inspector-reload.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript" src="inspector-helpers.js"></script>
  <script type="application/javascript">
"use strict";

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
};

let gInspectee = null;
let gWalker = null;
let gResourceWatcher = null;

addTest(async function setup() {
  const url = document.getElementById("inspectorContent").href;
  const { commands, doc } = await attachURL(url);
  const target = commands.targetCommand.targetFront;
  const inspector = await target.getFront("inspector");
  gInspectee = doc;
  const walker = inspector.walker;
  gWalker = await inspector.getWalker();
  gResourceWatcher = await createResourceWatcher(commands);

  ok(walker === gWalker, "getWalker() twice should return the same walker.");
  runNextTest();
});

addTest(async function testReload() {
  const oldRootID = gWalker.rootNode.actorID;

  info("Start watching for root nodes and wait for the initial root node");
  let rootNodeResolve;
  let rootNodePromise = new Promise(r => (rootNodeResolve = r));
  const onAvailable = rootNodeFront => rootNodeResolve(rootNodeFront);
  await gResourceWatcher.watchResources([gResourceWatcher.TYPES.ROOT_NODE], {
    onAvailable,
  });
  await rootNodePromise;

  info("Retrieve the node front for the selector `#a`");
  const nodeFront = await gWalker.querySelector(gWalker.rootNode, "#a");
  ok(nodeFront.actorID, "Node front has a valid actor ID");

  info("Reload the page and wait for the newRoot mutation");
  rootNodePromise = new Promise(r => (rootNodeResolve = r));

  gInspectee.defaultView.location.reload();
  await rootNodePromise;

  info("Retrieve the (new) node front for the selector `#a`");
  const newNodeFront = await gWalker.querySelector(gWalker.rootNode, "#a");
  ok(newNodeFront.actorID, "Got a new actor ID");
  ok(gWalker.rootNode.actorID != oldRootID, "Root node should have changed.");

  runNextTest();
});

addTest(function cleanup() {
  gWalker = null;
  gInspectee = null;
  gResourceWatcher = null;
  runNextTest();
});
  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
